--- src/pages/EsqueciSenha/index.js
@@ -1,1 +1,0 @@
-export { default } from "./EsqueciSenha";
