+++ src/routes.js
@@ -7,7 +7,6 @@ import Cadastro from "./pages/Cadastro";
 import Chat from "./pages/Chat";
 import CriacaoFormulario from "./pages/CriacaoFormulario";
 import EdicaoConteudo from "./pages/EdicaoConteudo";
-import EsqueciSenha from "./pages/EsqueciSenha";
 import Header from "./components/Header/Header";
 import ListaFormularios from "./pages/ListaFormularios";
 import ListaUsuarios from "./pages/ListaUsuarios";
